changed all longs to 128 bits

+++ include/generator.h
@@ -1,2 +1,2 @@
-unsigned long random_prime(int k);
+__uint128_t random_prime(int k);
 struct PrivateKey generate_keypair(int k);
