+++ primes.c
@@ -8,22 +8,22 @@
 #include <random.h>
 #include <primes.h>
 
-unsigned long eulers_totient(unsigned long p, unsigned long q) {
+__uint128_t eulers_totient(__uint128_t p, __uint128_t q) {
     return (p-1) * (q-1);
 }
 
-unsigned long carmicheals_totient(unsigned long p, unsigned long q) {
+__uint128_t carmicheals_totient(__uint128_t p, __uint128_t q) {
     return lcm(p-1, q-1);
 }
-unsigned long gcd(unsigned long a, unsigned long b) {
+__uint128_t gcd(__uint128_t a, __uint128_t b) {
     if (a && b) for(;(a %= b) && (b %= a););
     return a | b;
 }
 
-unsigned long lcm(unsigned long a, unsigned long b) {
-    long lcm;
-    long step;
-    long max = step =  a > b ? a : b;
+__uint128_t lcm(__uint128_t a, __uint128_t b) {
+    __int128_t lcm;
+    __int128_t step;
+    __int128_t max = step =  a > b ? a : b;
     for(;;) {
         if (max % a == 0 && max % b == 0) {
             lcm = max;
@@ -36,18 +36,18 @@ unsigned long lcm(unsigned long a, unsigned long b) {
 
 // This sometimes returns a different answer than the gcd function
 // which is a bug
-long extended_gcd(long a, long b, long* x, long* y) {
-    long old_r = a;
-    long r = b;
-    long old_s = 1;
-    long s = 0;
-    long old_t = 0;
-    long t = 1;
+__int128_t extended_gcd(__int128_t a, __int128_t b, __int128_t* x, __int128_t* y) {
+    __int128_t old_r = a;
+    __int128_t r = b;
+    __int128_t old_s = 1;
+    __int128_t s = 0;
+    __int128_t old_t = 0;
+    __int128_t t = 1;
 
     while (r != 0) {
-        long quotient = old_r / r;
+        __int128_t quotient = old_r / r;
         
-        long prov = r;
+        __int128_t prov = r;
         r = old_r - quotient * prov;
         old_r = prov;
         
@@ -112,20 +112,20 @@ unsigned int sieve_of_eratosthenes(unsigned int n, unsigned int** output) {
 
 }
 
-unsigned long power2(unsigned long exp) {
-    return (unsigned long) 1 << exp;
+__uint128_t power2(__uint128_t exp) {
+    return (__uint128_t) 1 << exp;
 }
 
-unsigned long power(unsigned long base, unsigned long exp) {
+__uint128_t power(__uint128_t base, __uint128_t exp) {
     if (base == 2) return power2(exp);
     int i;
-    unsigned long result = 1;
+    __uint128_t result = 1;
     for (i = 0; i < exp; i++) result *= base;
     return result;
 }
 
-unsigned long modpower2(unsigned long exp, unsigned long mod) {
-    unsigned long res = 1;
+__uint128_t modpower2(__uint128_t exp, __uint128_t mod) {
+    __uint128_t res = 1;
     for (;exp >= 1; exp--) {
         res = (res << 1) % mod;
     }
@@ -133,10 +133,10 @@ unsigned long modpower2(unsigned long exp, unsigned long mod) {
 }
 
 // Memory efficient modular exponentation with exponentation by squaring
-unsigned long modpower(unsigned long base, unsigned long exp, unsigned long mod) {
+__uint128_t modpower(__uint128_t base, __uint128_t exp, __uint128_t mod) {
     if (base == 2) return modpower2(exp,mod);
     int i;
-    unsigned long result = 1;
+    __uint128_t result = 1;
     while (exp > 0) {
         // Odd exponent
         if (exp & 1) {
@@ -150,7 +150,7 @@ unsigned long modpower(unsigned long base, unsigned long exp, unsigned long mod)
 
 //fails on 679
 //infinite loop 18446744072410345099
-bool isprime(unsigned long n) {
+bool isprime(__uint128_t n) {
     if (!(n&1) || !(n%3) || n <= 1) return false;
 
     if (n == 2 || n == 3) return true;
@@ -165,7 +165,7 @@ bool isprime(unsigned long n) {
     return true;
 }
 
-bool isprime_eratos(unsigned long n, unsigned int* first_primes, int amount) {
+bool isprime_eratos(__uint128_t n, unsigned int* first_primes, int amount) {
     for (int i=0; i<amount; i++) {
         if (n % first_primes[i] == 0) return false;
     }
@@ -175,11 +175,11 @@ bool isprime_eratos(unsigned long n, unsigned int* first_primes, int amount) {
 
 // This function is quite useless as it won't work for primes which create a large d
 // as this will cause an integer overflow if used as an exponent of a large base.
-bool miller_rabin(unsigned long n, int k) {    
+bool miller_rabin(__uint128_t n, int k) {    
     //write n as 2^s*d+1 where d is odd
-    unsigned long a;
-    unsigned long s = 0;
-    unsigned long d = n-1;
+    __uint128_t a;
+    __uint128_t s = 0;
+    __uint128_t d = n-1;
 
     if (!(n&1) || !(n%3) || n <= 1) return false;
 
@@ -194,7 +194,7 @@ bool miller_rabin(unsigned long n, int k) {
 
     for (int i=0; i<k; i++) {
         a = randlongrange(2,n-2); // testing 2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, and 37 should suffice for n < 2**64
-        unsigned long x = modpower(a, d, n);
+        __uint128_t x = modpower(a, d, n);
         if (x == 1 || x == n -1 ) continue;
         for (int j=0; j < s-1; j++) {
             x = x*x % n;
